pyproject.toml
[project]
name = "t"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pillow>=11.3.0",
]
